refactor(infra): split JsonlEventParser
