+++ src/JKToolKit.CodexSDK/Infrastructure/Internal/JsonlEventParsing/JsonlEventParserContext.cs
@@ -1,0 +1,6 @@
+using Microsoft.Extensions.Logging;
+
+namespace JKToolKit.CodexSDK.Infrastructure.Internal.JsonlEventParsing;
+
+internal readonly record struct JsonlEventParserContext(ILogger Logger);
+
